Add FTSO tech page
Contains diagrams and high level descriptions of all parts and processes.
It also has some technical details, with pointers to the relevant guides
in the docs (which are not great, currently).

Once the guides (about how to manually delegate, claim, etc) are up to par,
the instructions in the "Technical Details" section can be removed.